Shows app version at the bottom of Account switcher options Change package name from "app" to "carpe" to better read on running processes list
