+++ src-tauri/src/commands/mod.rs
@@ -6,6 +6,7 @@ mod tx;
 mod networks;
 mod query;
 mod examples;
+mod app_version;
 
 pub use wallets::*;
 pub use swarm::*;
@@ -15,3 +16,4 @@ pub use tx::*;
 pub use networks::*;
 pub use query::*;
 pub use examples::*;
+pub use app_version::*;
