+++ src-tauri/src/main.rs
@@ -12,7 +12,6 @@ pub mod configs_network;
 pub mod configs_profile;
 pub mod carpe_error;
 pub mod seed_peers;
-pub mod version;
 
 use tauri::{Menu, MenuItem, Submenu};
 use crate::{commands::*};
@@ -74,7 +73,7 @@ fn main() {
     set_env,
     debug_submit_proof_zero,
     // Version
-    get_version,
+    get_app_version,
  
     // Debug
     init_swarm,
